+++ DSW.HDWallet/Domain/Models/Wallet.cs
@@ -6,6 +6,6 @@
         public string? PublicKey { get; set; }
         public string? Path { get; set; }
         public int CoinIndex { get; set; }
-        public int? Balance { get; set; }
+        public long? Balance { get; set; }
     }
 }
